디버깅 참고 답안 완료

+++ 디버깅 연습 문제/Exercise_4/Knight.h
@@ -13,7 +13,11 @@ public:
 	bool IsDead();
 
 public:
-	int _hp;
+	// Before
+	// int _hp; 
+	
+	// After
+	long long _hp;
 	int _attack;
 };
